[refactor] orm config custom
