+++ backend/src/series/series.entity_.ts
