+++ backend/src/series/series.service.ts
@@ -1,4 +1,4 @@
-import { Series } from './series.entity';
+import { Series } from './series.entity_';
 import { Injectable } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Repository } from 'typeorm';
