--- backend/src/users/users.entity.ts
@@ -1,13 +1,0 @@
-import { Column, Entity, PrimaryGeneratedColumn } from 'typeorm';
-
-@Entity()
-export class Users {
-  @PrimaryGeneratedColumn()
-  id: number;
-
-  @Column({ unique: true })
-  userId: string;
-
-  @Column()
-  password: string;
-}
