+++ backend/src/users/users.module.ts
@@ -2,7 +2,7 @@ import { Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
 
 import { UsersService } from '@users/users.service';
-import { Users } from './users.entity';
+import { Users } from '@entity/Users';
 
 @Module({
   // imports: [TypeOrmModule.forFeature([Users])],
